requirements.txt
arrow==1.3.0
asgiref==3.7.2
binaryornot==0.4.4
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.39.0
cookiecutter==2.6.0
dj-database-url==2.1.0
Django==5.0.2
django-cloudinary-storage==0.3.0
django-extensions==3.2.3
django-tailwind==3.8.0
djangorestframework==3.14.0
easyocr==1.7.1
filelock==3.13.1
fsspec==2024.2.0
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-vision==3.7.1
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
gunicorn==21.2.0
idna==3.6
imageio==2.34.0
Jinja2==3.1.3
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==23.2
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
Pygments==2.17.2
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
rsa==4.9
scikit-image==0.22.0
scipy==1.12.0
shapely==2.0.3
six==1.16.0
sqlparse==0.4.4
sympy==1.12
text-unidecode==1.3
tifffile==2024.2.12
torch==2.2.1
torchvision==0.17.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
urllib3==2.2.1
